move store under vcs
