+++ packages/insomnia/src/sync/vcs/store/hooks/compress.ts
