+++ packages/insomnia/src/ui/components/modals/sync-delete-modal.tsx
@@ -3,7 +3,7 @@ import { OverlayContainer } from 'react-aria';
 
 import { strings } from '../../../common/strings';
 import { useWorkspaceLoaderData } from '../../../routes/organization.$organizationId.project.$projectId.workspace.$workspaceId';
-import { interceptAccessError } from '../../../sync/vcs/util';
+import { interceptAccessError } from '../../../sync/access-error';
 import { Button } from '../../components/themed-button';
 import { Modal, type ModalHandle, type ModalProps } from '../base/modal';
 import { ModalBody } from '../base/modal-body';
